+++ extensions/cursor-directory/CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## [Features & Improvements] - 2024-10-12
 
-- Remove integration with cursor-recent-projects extension
-- Add a new command to easily open project and automatically apply rule in Cursor
+- Add a new command to easily open project and automatically apply rule in Cursor (if you have configured the default `Projects Directory`)
 - Improve error handling
 - Update dependencies